Manage processes in an environment
Turns out it's not the performance overhead of shared_ptrs that's a
problem, it's the extra lifecycle management that you have to think
about.  Instead, we'll copy what the C library does, and have an
"environment" object that owns all of the processes that you create.  It
will also ensure that there's a single canonical copy of each process.