Calculate τ closures
We'll need to calculate these a lot during normalization and refinement.
This is similar to the operation that translates an NFA into a DFA when
working with regular expressions.  Given a set of processes, we need to
extend that set with any additional processes that you could get to by
only following hidden events (which are called τ in CSP).