Add prenormalized processes
This is the first step in a refinement check.  It's equivalent to
transforming an NFA to a DFA when working with regular expressions.
Instead of our LTS graph nodes representing individual processes, each
node represents a set of processes; we merge together all of the states
that are reachable by following only hidden τ events; after having
followed a particular trace of visible events, you might be an any of
those states.

We implement this using the same Process interface as our "normal" CSP
operators; the only difference is that we guarantee that `initials` will
never return τ, and that `afters` will never return more than one
outgoing process for any event.