Add traces behavior
This is the first semantic model that we want to support.  Each semantic
model will have its own wrapper struct, containing all of the classes
and methods that define that method.  That will let us parameterize
certain bits of the refinement checker in a nicely generic way.

In the traces model, the behavior of a process is the set of non-τ
events that it can perform.  That's easy enough to calculate, just by
calling the process's `initials` method.