Add normalized processes
This is like a prenormalized process, but we also perform a bisimulation
to find all of the prenormalized subprocesses that have equivalent
behavior.  The results of that bisimulation might be different depending
on which "root" node we start the normalization from, so we have to
include the root in the "identity" of the normalized process.  That
mostly only has an effect on the CSP₀ parser, and really, you should
only use CSP₀'s `normalize` operator in test cases.