Include all subprocesses when printing recursion
Before we used a behavioral definition of reachability to find all of
the recursive definitions to include when rendering a `let` statement.
Some operators (e.g. external choice) can pull apart their definitions
when producing initials and afters sets, which could "hide" some of the
recursive definitions, giving us an incomplete `let` statement.  We
actually want to search through the *syntactic* subprocesses, not the
*behavioral* ones, when constructing a process's definition.