Add refinement checker
Now that all of the pieces are in place, we just have to do a BFS
through pairs of Spec and Impl processes.  If we find any pair where the
Impl subprocess's behavior doesn't refine the Spec subprocess's
behavior, the refinement check fails.  If we make it through to the end
without finding a bad pair, the check succeeds.