Fix hashing and comparison of bags and sets
I was being sloppy with the hashing of bags and sets, using a bare XOR
when I should really be mixing in the elements properly.  It's not as
much of a problem for sets, since we're guaranteed that each element
will appear at most once (so the `x ^ x == 0` problem can't happen).
With bags it's a real problem, though.  But if I'm going to fix one, I
might as well fix both.