Better representation for process bags
The STL version seems to store each copy of each process individually,
and each element of the map seems to require its own memory allocation.
That's a lot of memory churn, especially when you insert and remove
elements from a bag in quick succession, just like we do when
calculating the afters of an interleave process.